Added authentication

+++ RestaurantAPI/Services/RestaurantService.cs
@@ -1,5 +1,8 @@
-﻿using AutoMapper;
+﻿using System.Security.Claims;
+using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using RestaurantAPI.Authorization;
 using RestaurantAPI.Entities;
 using RestaurantAPI.Exceptions;
 using RestaurantAPI.Models;
@@ -20,11 +23,21 @@ namespace RestaurantAPI.Services
         private readonly RestaurantDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger _logger; 
-        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger)
+        private readonly IAuthorizationService _authorizationService;
+        private readonly IUserContextService _userContextService;
+        public RestaurantService
+        (RestaurantDbContext dbContext,
+            IMapper mapper, 
+            ILogger<RestaurantService> logger,
+            IAuthorizationService authorizationService,
+            IUserContextService userContextService
+            )
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
+            _userContextService = userContextService;
         }
 
         public void Update(int id, UpdateRestaurantDto dto)
@@ -34,6 +47,14 @@ namespace RestaurantAPI.Services
                 .FirstOrDefault(r => r.Id ==id);
 
             if (restaurant == null) throw new NotFoundException("Restaurant not found");
+
+            var authorizationResult = _authorizationService
+                .AuthorizeAsync(_userContextService.User, restaurant, new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             restaurant.Name = dto.Name;
             restaurant.Description = dto.Description;
@@ -51,6 +72,14 @@ namespace RestaurantAPI.Services
                 .FirstOrDefault(r => r.Id == id);
 
             if (restaurant == null) throw new NotFoundException("Restaurant not found");
+
+            var authorizationResult = _authorizationService
+                .AuthorizeAsync(_userContextService.User, restaurant, new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             _dbContext.Restaurants.Remove(restaurant);
             _dbContext.SaveChanges();
@@ -86,6 +115,7 @@ namespace RestaurantAPI.Services
         public int Create(CreateRestaurantDto dto)
         {
             var restaurant = _mapper.Map<Restaurant>(dto);
+            restaurant.CreatedById = _userContextService.GetUserId;
             _dbContext.Restaurants.Add(restaurant);
             _dbContext.SaveChanges();
